使用extendMessageConverters扩展gson转换器
